fix: 데몬어벤져 세트 HP 는 줄마다 내림, 보정 상한은 float 0.85
상위 캐릭터 전원이 -12.7ppm 어긋나 보정 상한을 0.8499892 로 두었던 것은 세트 효과
HP 를 합계로 절반 내던 오차였다. 칠흑 세트 5세트부터 붙는 "최대 HP +375" 줄이 줄마다
187 로 내림되고, 그래서 칠흑을 채운 캐릭터일수록(=상위) 결손이 커졌다. 줄 단위로
내리자 상한 구간 잔차 중앙이 -12.5 → +0.03ppm 이 되고 상한은 0.85 로 돌아왔다.

남은 -45(흡혈 16.9억 기준)는 0.85 를 float 로 든 값(0.8500000238)이었다. 상·하한을
float 로 캐스팅하니 흡혈 75일 패널 72개 상태가 전부 정수 일치한다.

총 HP 정수화는 반올림이 아니라 내림이다(2,894명에서 29 대 15).

09-11 표본 3,804명: 정수 일치 26 → 229, 전투력 ±10 안 193 → 620, ±100 안 1,695 → 2,763,
상대오차 1e-6 안 1,929 → 3,100(81.5%). 골든 데벤 10명은 ±0.5ppm, 흡혈은 0.

+++ src/main/java/org/whitedoggy/mapleweb2/domain/common/stat/GameData.java
@@ -101,12 +101,17 @@ public record GameData(
             return conversionHpPerStar.isEmpty() ? 0 : stars * conversionHpPerStar.getLast().hp();
         }
 
+        /**
+         * 상·하한은 게임이 단정밀도(float)로 들고 있다 — 0.85 는 0.8500000238 이다. 상한 구간
+         * 166명 중 128명이 이 값으로 정수 일치하고, 배정밀도 0.85 로는 전원이 -3 이하로 어긋난다
+         * (2026-09-15). 직선 구간은 배정밀도가 더 잘 맞는다.
+         */
         public double correctionOf(double base) {
             if (base <= 0) {
-                return correction.min();
+                return (float) correction.min();
             }
             double raw = correction.intercept() + correction.slope() * Math.log10(base);
-            return Math.max(correction.min(), Math.min(correction.max(), raw));
+            return Math.max((float) correction.min(), Math.min((float) correction.max(), raw));
         }
     }
